UserInfoAuthoritiesService renamed to UserInfoAuthoritiesRetriever
